+++ google2/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file documents changes to the `workbench-google2` library, including notes on how to upgrade to new versions.
 
+## 0.7
+Changed
+- Use `io.chrisdavenport.log4cats.StructuredLogger` instead of `io.chrisdavenport.log4cats.Logger`
+
+Add
+- Add `tracedPublish` to `GooglePublisher[F]`
+- Log messages with traceId in `GoogleSubscriberInterpreter`
+
+SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.7-TRAVIS-REPLACE-ME"`
+
 ## 0.6
 Changed
 - Bump `fs2-io` to `2.0.1`
@@ -13,12 +23,9 @@ Changed
 
 Add
 - Add `GoogleDataproc` and `GoogleDataprocInterpreter`
-- Add `tracedPublish` to `GooglePublisher[F]`
-- Log messages with traceId in `GoogleSubscriberInterpreter`
 - Add `delete` function to `GoogleTopicAdmin` trait and implementation
 
 SBT dependency: `"org.broadinstitute.dsde.workbench" %% "workbench-google2" % "0.6-c91d96b"`
-
 
 ## 0.5